feat: split up base class
